creation of folders to separate the view from the viewmodel and login mock
