+++ src/frontend/lib/ViewModel/loginViewModel.dart
@@ -1,0 +1,9 @@
+class LoginViewModel {
+
+  bool logIn(String email, String password){
+    if(email == "aaa" && password == "111")
+      return true;
+    
+    return false;
+  }
+}
